seperate auth with user.controller.
